fullstack-for-fe/compilation/backend/requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
-e git+https://github.com/ryanzhou7/dev-notes.git@36200d1bdd70d676a7faf9a83736b3337541fe20#egg=flaskr&subdirectory=fullstack-for-fe/compilation/backend
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.41
Werkzeug==2.2.2
